% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%fin\%}
\alias{\%fin\%}
\title{Fast replacment for the %in% operator}
\usage{
x \%fin\% table
}
\arguments{
\item{x}{Values to be matched.}

\item{table}{Values to be matched against.}
}
\value{
A logical vector with the same length as \code{x}. It is \code{TRUE} 
wherever an element of \code{x} is a member of \code{table}.
wherever an element of \code{x} is a member of \code{table}.
}

